Tweaks in ED table and addition of PUSH/POP RP and NCALL a16/HL
Added a difference between undefined and unimplemented opcode - undefined
opcodes throw a warning and afterwards, just ignored, and unimplemented
opcodes terminate the emulator.
PUSH/POP RP are up for working with XY, just XY dispatches are not yet
designed and implemented.
NCALL is coded as ED D0/D1 (a16 and HL) and does a call to native code.